fix impl

+++ SuperDodawanie/src/main/java/pl/edu/amu/wmi/bikol/dino/superdodawanie/Dodawanie.java
@@ -12,7 +12,7 @@ package pl.edu.amu.wmi.bikol.dino.superdodawanie;
 public class Dodawanie {
     
     public String dodaj(String a, String b){
-        return Integer.toString(Integer.parseInt(a) + Integer.parseInt(b));
+        return Double.toString(Double.parseDouble(a) + Double.parseDouble(b));
     }
     
 }
